BankMGMT/requirements.txt
pandas
Pillow
mysql-connector
datetime
ttkthemes
tkcalendar
matplotlib